4-Note Sequence Permutations
This program calculates the number of 4 note sequences
which are possible if the first note is fixed

The remaining three notes can be any of the 12 notes in an octave. (These notes are A, A#, B, C, C#, D, D#, E, F, F#, G, G#).
As is music, repetition is allowed. The same note can be used multiple times. Order is also important. The program calculates permutations and outputs the number of possible sequences